fix data type

+++ data/common_crawl/prepare.py
@@ -43,7 +43,7 @@ tokenized = split_dataset.map(
 for split, dset in tokenized.items():
     arr_len = np.sum(dset['len'])
     filename = os.path.join(os.path.dirname(__file__), f'{split}.bin')
-    dtype = np.uint16  # (can do since enc.max_token_value == 50256 is < 2**16)
+    dtype = np.uint32  # well, we have so much tokens...
     arr = np.memmap(filename, dtype=dtype, mode='w+', shape=(arr_len,))
 
     print(f"writing {filename}...")
